uint 64!

+++ main/cmds.go
@@ -127,7 +127,7 @@ func enablePre(ctx *log.Context, hEnv HandlerEnvironment, seqNum int) error {
 		utils.TryClearExtensionScriptsDirectoriesAndSettingsFilesExceptMostRecent(hEnv.HandlerEnvironment.LogFolder,
 			hEnv.HandlerEnvironment.ConfigFolder,
 			"customscriptextension",
-			uint(seqNum),
+			uint64(seqNum),
 			"\\d+.settings",
 			"%d.settings")
 		//settings.CleanUpSettings(el, hEnv.HandlerEnvironment.ConfigFolder)
@@ -183,7 +183,7 @@ func enable(ctx *log.Context, h HandlerEnvironment, seqNum int) (string, error) 
 	utils.TryClearExtensionScriptsDirectoriesAndSettingsFilesExceptMostRecent(h.HandlerEnvironment.LogFolder,
 		h.HandlerEnvironment.ConfigFolder,
 		"customscriptextension",
-		uint(seqNum),
+		uint64(seqNum),
 		"\\d+.settings",
 		"%d.settings")
